% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/au.population.change.R
\docType{data}
\name{au.population.change}
\alias{au.population.change}
\title{All population data for Australia}
\format{Data frame with columns
\describe{
 \item{location}{State abbreviation or 'Australia'}
 \item{year}{Year}
 \item{population}{Population}
 \item{sex}{Male, Female, Persons (all data)}
}}
\source{
ABS 3105.0.65.0012014 Table 1.3
}
\usage{
au.population.change
}
\description{
All population change data for Australia from 1971-2011 by state and nationally - including births, deaths and international and interstate migration.
}
\examples{
au.population.change
}
\keyword{datasets}

