% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/au.population.recent.R
\docType{data}
\name{au.population.recent}
\alias{au.population.recent}
\title{All population data for Australia}
\format{Data frame with columns
\describe{
 \item{location}{State abbreviation or 'Australia'}
 \item{year}{Year}
 \item{population}{Population}
 \item{sex}{Male, Female, Persons (all data)}
}}
\source{
ABS 3105.0.65.0012014 Table 1.3
}
\usage{
au.population.recent
}
\description{
All population data for Australia from 1901-2012 by state and nationally.
}
\examples{
au.population.recent
}
\keyword{datasets}

